+++ packages/config/src/mutations/update.js
@@ -16,7 +16,7 @@ import { applyMutations } from './apply.js'
 // If `netlify.toml` does not exist, creates it. Otherwise, merges the changes.
 export const updateConfig = async function (
   configMutations,
-  { buildDir, configPath, headersPath, redirectsPath, context, branch, logs },
+  { buildDir, configPath, headersPath, redirectsPath, context, branch, logs, featureFlags },
 ) {
   if (configMutations.length === 0) {
     return
@@ -25,8 +25,8 @@ export const updateConfig = async function (
   const inlineConfig = applyMutations({}, configMutations)
   const normalizedInlineConfig = ensureConfigPriority(inlineConfig, context, branch)
   const updatedConfig = await mergeWithConfig(normalizedInlineConfig, configPath)
-  const configWithHeaders = await addHeaders(updatedConfig, headersPath, logs)
-  const finalConfig = await addRedirects(configWithHeaders, redirectsPath, logs)
+  const configWithHeaders = await addHeaders({ config: updatedConfig, headersPath, logs, featureFlags })
+  const finalConfig = await addRedirects({ config: configWithHeaders, redirectsPath, logs, featureFlags })
   const simplifiedConfig = simplifyConfig(finalConfig)
   await backupConfig({ buildDir, configPath, headersPath, redirectsPath })
   await Promise.all([